Many code design changes.
Started abstracting code into better abstractions. Code is now
basically separated into two pieces - Server; Client. Because of this
most of the namespaces are now changed. Server deals with code parsing
and execution and can be ran on a separate thread (not possible yet
though). Client deals with getting user input. The idea behind this is
to allow Web based user input where Server would be running on a
separate thread. Almost none of the AIP functions work on this commit
so it isn't recommended to use this feature branch.